Revert "Feature/quiz progress"
