+++ components/LoadingSpinner.tsx
@@ -4,7 +4,6 @@ export function LoadingSpinner() {
   return (
     <div className='flex items-center justify-center min-h-screen'>
       <Loader2 className='h-8 w-8 animate-spin' />
-      <p className='mt-4'>Loading...</p>
     </div>
   )
 }
