+++ components/ProtectedRoute.tsx
@@ -2,7 +2,7 @@
 import { useAuth } from '@/contexts/AuthContext'
 import { useRouter } from 'next/navigation'
 import { useEffect } from 'react'
-import { LoadingSpinner } from '@/components/LoadingSpinner'
+import { Loader2 } from 'lucide-react'
 
 interface ProtectedRouteProps {
   children: React.ReactNode
@@ -19,7 +19,11 @@ export default function ProtectedRoute({ children }: ProtectedRouteProps) {
   }, [user, loading, router])
 
   if (loading) {
-    return <LoadingSpinner />
+    return (
+      <div className='flex items-center justify-center min-h-screen'>
+        <Loader2 className='w-8 h-8 animate-spin' />
+      </div>
+    )
   }
 
   if (!user) return null
