+++ components/classrooms/CreateClassroomForm.tsx
@@ -30,9 +30,8 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select'
-import { LoadingSpinner } from '@/components/LoadingSpinner'
-import { useToast } from '@/hooks/use-toast'
 import { Loader2 } from 'lucide-react'
+import { useToast } from '@/hooks/use-toast'
 
 interface Props {
   teacherId: string
@@ -201,7 +200,13 @@ export function CreateClassroomForm({ teacherId, teacherSchool }: Props) {
     })
   }
 
-  if (loading) return <LoadingSpinner />
+  if (loading) {
+    return (
+      <div className='flex items-center justify-center min-h-[400px]'>
+        <Loader2 className='h-8 w-8 animate-spin' />
+      </div>
+    )
+  }
 
   return (
     <Form {...form}>
