+++ components/ui/responsive-search.tsx
@@ -62,12 +62,15 @@ export function ResponsiveSearch() {
   return (
     <div>
       <Button
-        variant='ghost'
-        // size='sm'
-        // className='w-10 px-0'
+        variant='slate'
+        className='relative h-9 w-9 p-0 xl:h-10 xl:w-60 xl:justify-start xl:px-3 xl:py-2'
         onClick={() => setOpen(true)}
       >
-        <Search className='h-6 w-6 xl:mr-2' />
+        <Search className='h-4 w-4 xl:mr-2' />
+        <span className='hidden xl:inline-flex'>Search content...</span>
+        <kbd className='pointer-events-none absolute right-1.5 top-2 hidden h-6 select-none items-center gap-1 rounded border bg-muted px-1.5 font-mono text-xs font-normal opacity-100 xl:flex'>
+          <span className='text-xs'>⌘</span>K
+        </kbd>
       </Button>
       <CommandDialog
         open={open}
