+++ types/quiz/quiz.ts
@@ -15,25 +15,3 @@ export interface Quiz {
   imageUrl: string
   questions: Question[]
 }
-
-export interface QuizProgress {
-  passed: boolean
-  lastTaken: number | null
-  attempts: number
-  successfulAttempts: number
-  highestScore: number
-  score: number
-  correctAnswers: number
-  totalQuestions: number
-  selectedAnswers: number[]
-}
-
-export interface QuizAttempt {
-  id: string
-  timestamp: number
-  score: number
-  correctAnswers: number
-  totalQuestions: number
-  passed: boolean
-  selectedAnswers: number[]
-}
